JBIDE-13624 BrowserSim: source code refactoring
refactoring
