+++ plugins/org.jboss.tools.vpe.browsersim/src/org/jboss/tools/vpe/browsersim/util/BrowserSimUtil.java
@@ -1,0 +1,115 @@
+/*******************************************************************************
+ * Copyright (c) 2007-2013 Red Hat, Inc.
+ * Distributed under license by Red Hat, Inc. All rights reserved.
+ * This program is made available under the terms of the
+ * Eclipse Public License v1.0 which accompanies this distribution,
+ * and is available at http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributor:
+ *     Red Hat, Inc. - initial API and implementation
+ ******************************************************************************/
+package org.jboss.tools.vpe.browsersim.util;
+
+import java.text.MessageFormat;
+
+import org.eclipse.swt.events.DisposeEvent;
+import org.eclipse.swt.events.DisposeListener;
+import org.eclipse.swt.graphics.Point;
+import org.eclipse.swt.graphics.Rectangle;
+import org.eclipse.swt.graphics.Resource;
+import org.eclipse.swt.widgets.Monitor;
+import org.eclipse.swt.widgets.Shell;
+import org.eclipse.swt.widgets.Widget;
+import org.jboss.tools.vpe.browsersim.model.Device;
+import org.jboss.tools.vpe.browsersim.model.SkinMap;
+import org.jboss.tools.vpe.browsersim.ui.MessageBoxWithLinks;
+import org.jboss.tools.vpe.browsersim.ui.Messages;
+import org.jboss.tools.vpe.browsersim.ui.skin.BrowserSimSkin;
+
+/**
+ * @author Yahor Radtsevich (yradtsevich)
+ * @author Konstantin Marmalyukov (kmarmaliykov)
+ */
+
+public class BrowserSimUtil {
+	private static final String ABOUT_ICON = "icons/browsersim_32px.png"; //$NON-NLS-1$
+	
+	public static void fixShellLocation(Shell shell, Rectangle clientArea) {
+		Point shellLocation = shell.getLocation();
+		Point shellSize = shell.getSize();
+		int bottomOverlap = shellLocation.y + shellSize.y - (clientArea.y + clientArea.height);
+		if (bottomOverlap > 0) {
+			if (shellLocation.y > bottomOverlap) {
+				shellLocation.y -= bottomOverlap;
+			} else {
+				shellLocation.y = 0;
+			}
+		}
+
+		int rightOverlap = shellLocation.x + shellSize.x - (clientArea.x + clientArea.width);
+		if (rightOverlap > 0) {
+			if (shellLocation.x > rightOverlap) {
+				shellLocation.x -= rightOverlap;
+			} else {
+				shellLocation.x = 0;
+			}
+		}
+
+		shell.setLocation(shellLocation);
+	}
+	
+	public static Rectangle getMonitorClientArea(Monitor monitor) {
+		Rectangle clientArea = monitor.getClientArea();
+
+		/* on Linux returned monitor client area may be bigger
+		 * than the monitor bounds when multiple monitors are used.
+		 * The following code fixes this */
+		Rectangle bounds = monitor.getBounds();
+		clientArea.width = Math.min(clientArea.width, bounds.width);
+		clientArea.height = Math.min(clientArea.height, bounds.height);
+
+		return clientArea;
+	}
+	
+	/**
+	 * See JBIDE-11896	BrowserSim: pixel ratio problem.
+	 * 
+	 * On many mobile devices like iPhone 4 1 CSS pixel = 2 device pixels.
+	 */
+	public static Point getSizeInDesktopPixels(Device device) {
+		double pixelRatio = device.getPixelRatio();
+		if (device.getPixelRatio() == 0.0) {
+			pixelRatio = 1.0;
+			new RuntimeException("Pixel Ratio is 0.0").printStackTrace();
+		}
+		int width = (int) Math.round(device.getWidth() / pixelRatio);
+		int height = (int) Math.round(device.getHeight() / pixelRatio);
+		return new Point(width, height);
+	}
+
+	public static Class<? extends BrowserSimSkin> getSkinClass(Device device, boolean useSkins) {
+		return SkinMap.getInstance().getSkinClass(useSkins ? device.getSkinId() : null);
+	}
+	
+
+	public static void showAboutDialog(Shell shell) {
+		String message;
+		String version = ManifestUtil.getManifestVersion();
+		if (version != null) {
+			message = MessageFormat.format(Messages.BrowserSim_ABOUT_MESSAGE, ManifestUtil.getManifestVersion());
+		} else {
+			message = MessageFormat.format(Messages.BrowserSim_ABOUT_MESSAGE, ""); //$NON-NLS-1$
+		}
+		ImageList imageList = new ImageList(shell);
+		new MessageBoxWithLinks(shell, message, imageList.getImage(ABOUT_ICON), Messages.BrowserSim_ABOUT).open();
+	}
+	
+	public static void addDisposeListener(Widget widget, final Resource disposable) {
+		widget.addDisposeListener(new DisposeListener() {
+			@Override
+			public void widgetDisposed(DisposeEvent e) {
+				disposable.dispose();
+			}
+		});
+	}
+}
